Match data types of IDs and foreign keys

+++ database/migrations/2023_10_23_192616_create_courses_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->boolean('is_public');
             $table->timestamps();
 
-            $table->integer('owner');
+            $table->unsignedBigInteger('owner');
             $table->foreign('owner')->references('id')->on('users');
         });
     }
